added redirect + moved to market
